sm8250: Import G9810ZCU2BTH6 from OSRC

Conflicts:
	Documentation/arm64/silicon-errata.txt
	arch/arm64/Kconfig
	arch/arm64/Kconfig.projects
	arch/arm64/Kconfig.xyz
	arch/arm64/configs/vendor/x1q_chn_openx_defconfig
	arch/arm64/include/asm/cpucaps.h
	arch/arm64/include/asm/cputype.h
	arch/arm64/kernel/cpu_errata.c
	arch/arm64/kernel/cpufeature.c
	arch/arm64/kernel/smp.c
	drivers/bluetooth/bluetooth-power.c
	drivers/bluetooth/btfm_slim.c
	drivers/bus/mhi/controllers/mhi_arch_qcom.c
	drivers/bus/mhi/controllers/mhi_qcom.c
	drivers/bus/mhi/core/mhi_init.c
	drivers/bus/mhi/core/mhi_internal.h
	drivers/bus/mhi/core/mhi_main.c
	drivers/bus/mhi/core/mhi_pm.c
	drivers/bus/mhi/devices/mhi_netdev.c
	drivers/clk/clk.c
	drivers/clk/qcom/camcc-lagoon.c
	drivers/clk/qcom/clk-rpmh.c
	drivers/clk/qcom/debugcc-lagoon.c
	drivers/clk/qcom/dispcc-lagoon.c
	drivers/clk/qcom/gpucc-lagoon.c
	drivers/clk/qcom/npucc-lagoon.c
	drivers/cpufreq/qcom-cpufreq-hw.c
	drivers/gpu/drm/drm_edid.c
	drivers/gpu/msm/adreno-gpulist.h
	drivers/hid/hid-quirks.c
	drivers/hid/hid-qvr.c
	drivers/i2c/busses/i2c-qcom-geni.c
	drivers/input/input.c
	drivers/input/touchscreen/Makefile
	drivers/input/touchscreen/focaltech_touch/focaltech_common.h
	drivers/input/touchscreen/focaltech_touch/focaltech_core.c
	drivers/input/touchscreen/focaltech_touch/focaltech_core.h
	drivers/input/touchscreen/focaltech_touch/focaltech_ex_mode.c
	drivers/input/touchscreen/focaltech_touch/focaltech_flash.c
	drivers/input/touchscreen/focaltech_touch/focaltech_flash/Makefile
	drivers/input/touchscreen/stm/fts9cu80f_b/fts_ts.c
	drivers/iommu/arm-smmu.c
	drivers/iommu/iova.c
	drivers/media/platform/msm/cvp/cvp_hfi_helper.h
	drivers/media/platform/msm/cvp/msm_cvp.c
	drivers/media/platform/msm/cvp/msm_cvp_dsp.c
	drivers/media/platform/msm/dvb/demux/mpq_dmx_plugin_common.c
	drivers/media/platform/msm/npu/npu_dev.c
	drivers/media/platform/msm/npu/npu_mgr.c
	drivers/media/platform/msm/synx/synx.c
	drivers/misc/Kconfig
	drivers/misc/Makefile
	drivers/net/wireless/broadcom/bcmdhd_101_12/dhd_common.c
	drivers/net/wireless/broadcom/bcmdhd_101_12/wl_android.c
	drivers/net/wireless/cnss2/Kconfig
	drivers/net/wireless/cnss2/Makefile
	drivers/net/wireless/cnss2/bus.c
	drivers/net/wireless/cnss2/bus.h
	drivers/net/wireless/cnss2/main.c
	drivers/net/wireless/cnss2/main.h
	drivers/net/wireless/cnss2/pci.c
	drivers/net/wireless/cnss2/pci.h
	drivers/net/wireless/cnss2/qmi.c
	drivers/net/wireless/cnss_utils/wlan_firmware_service_v01.c
	drivers/net/wireless/cnss_utils/wlan_firmware_service_v01.h
	drivers/pci/controller/pci-msm.c
	drivers/perf_mgr/perf_mgr.c
	drivers/platform/msm/gsi/gsi.c
	drivers/platform/msm/ipa/ipa_api.c
	drivers/platform/msm/ipa/ipa_clients/ipa_wdi3.c
	drivers/platform/msm/ipa/ipa_v3/ipa.c
	drivers/platform/msm/ipa/ipa_v3/ipa_i.h
	drivers/platform/msm/ipa/ipa_v3/ipa_mpm.c
	drivers/platform/msm/ipa/ipa_v3/ipa_nat.c
	drivers/platform/msm/ipa/ipa_v3/ipa_rt.c
	drivers/platform/msm/ipa/ipa_v3/rmnet_ipa.c
	drivers/power/supply/qcom/battery.c
	drivers/power/supply/qcom/qg-soc.c
	drivers/power/supply/qcom/qg-util.c
	drivers/power/supply/qcom/qpnp-smb5.c
	drivers/power/supply/qcom/smb1398-charger.c
	drivers/scsi/ufs/ufs-sysfs.c
	drivers/scsi/ufs/ufs.h
	drivers/scsi/ufs/ufshcd.c
	drivers/soc/qcom/dcc_v2.c
	drivers/soc/qcom/dfc_qmi.c
	drivers/soc/qcom/fsa4480-i2c.c
	drivers/soc/qcom/llcc-lagoon.c
	drivers/soc/qcom/qmi_rmnet.c
	drivers/soc/qcom/rpmh.c
	drivers/soc/qcom/smcinvoke.c
	drivers/spi/spi-geni-qcom.c
	drivers/tty/serial/msm_geni_serial.c
	drivers/usb/dwc3/dwc3-msm.c
	drivers/usb/dwc3/gadget.c
	drivers/usb/phy/phy-msm-qusb-v2.c
	fs/crypto/fscrypt_ddar.c
	fs/crypto/fscrypt_knox_private.h
	fs/crypto/keyinfo.c
	fs/crypto/sdp/sdp_dek.c
	fs/ext4/ioctl.c
	fs/f2fs/checkpoint.c
	fs/f2fs/data.c
	fs/f2fs/f2fs.h
	fs/f2fs/file.c
	fs/f2fs/gc.c
	fs/f2fs/segment.c
	fs/f2fs/segment.h
	fs/f2fs/super.c
	include/asm-generic/qspinlock.h
	include/dt-bindings/clock/mdss-7nm-pll-clk.h
	include/linux/fscrypt.h
	include/linux/input/input_booster.h
	include/linux/mhi.h
	include/linux/qcom-geni-se.h
	include/linux/sched/sysctl.h
	include/net/cnss2.h
	include/net/flow_dissector.h
	include/trace/events/dfc.h
	include/uapi/linux/fs.h
	include/uapi/linux/nl80211.h
	kernel/sched/core.c
	kernel/sched/fair.c
	kernel/sched/sched.h
	kernel/sched/walt.c
	net/wireless/chan.c
	net/wireless/util.c
	security/selinux/hooks.c
	sound/usb/usb_audio_qmi_svc.c
	techpack/video/msm/vidc/hfi_common.c
	techpack/video/msm/vidc/msm_venc.c
	techpack/video/msm/vidc/msm_vidc_common.c
	techpack/video/msm/vidc/msm_vidc_platform.c

+++ drivers/scsi/scsi_lib.c
@@ -1554,7 +1554,7 @@ void scsi_alloc_tw(struct scsi_device *sdev)
 		blk_register_tw_try_on_fn(sdev->request_queue, scsi_tw_try_on_fn);
 		blk_register_tw_try_off_fn(sdev->request_queue, scsi_tw_try_off_fn);
 		printk(KERN_INFO "%s: register scsi ufs tw interface for LU %d\n",
-				__func__, sdev->lun);
+				__func__, (int)sdev->lun);
 	}
 }
 #endif
